Updated the RetHat default params.
Removed '/usr/bin/GET' from the default scriptwhitelist parameter for
RedHat as the script is not a valid path on normal installs.

This entry is found in the CentOS install and seems to work fine when
the OS supplied config is used. However, when the cleaned out install
that this module creates is used the config check fails.  It appears to
be an error in how rkhunter interprets the rkhunter.conf.